add tokens to token factory genesis and move Ticker and CurrencyId types to token factory
